+++ src/components/features/user/ui/component/follow/UserRow.tsx
@@ -1,7 +1,7 @@
 import Avatar from "@/components/ui/avatar/Avatar";
 import type { UserRowProps } from "../../../types";
 import FollowButton from "@/components/ui/button/FollowButton";
-import { useAddFollow, useIsConnected } from "../../../hooks";
+import { useAddFollow, useIsConnected } from "@/shared/hooks/user";
 
 export default function UserRow({ user, isMe }: UserRowProps) {
   const { isConnected } = useIsConnected(user.id);
