+++ src/components/features/user/ui/component/profile/ProfileActions.tsx
@@ -1,8 +1,8 @@
 import { Button } from "@/components/ui/button/Button";
 import { useRouter } from "next/router";
 import FollowButton from "@/components/ui/button/FollowButton";
-import { useAddFollow, useIsConnected } from "../../../hooks";
 import { ProfileActionsProps } from "../../../types";
+import { useAddFollow, useIsConnected } from "@/shared/hooks/user";
 
 export default function ProfileActions({ isMe, userId }: ProfileActionsProps) {
   const router = useRouter();
